Release v25.08 Beta
* Add Metal backend support for macOS.
* Update dependency and build system to support cross-platform development.
* Fix usdview crash [Part2]: update slang.
* Fix crash when loading AutodeskTelescope using usdview [Part1].
  * Fix out-of-bound crashes
  * Fix normal_image MaterialX definitions.
  * Fix warning when getting data from no empty VtValues.
* Fix build error on latest VS2022.
* Fix runtime folder plugin files.
* Fix Loading Large Size Image Crash with HgiRenderer.
* Fix Image Resolver Plugin.

+++ Libraries/Aurora/Source/MaterialDefinition.cpp
@@ -1,12 +1,16 @@
+// Copyright 2025 Autodesk, Inc.
 //
-// Copyright 2023 by Autodesk, Inc.  All rights reserved.
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
 //
-// This computer source code and related instructions and comments
-// are the unpublished confidential and proprietary information of
-// Autodesk, Inc. and are protected under applicable copyright and
-// trade secret law.  They may not be disclosed to, copied or used
-// by any third party without the prior written consent of Autodesk, Inc.
+// http://www.apache.org/licenses/LICENSE-2.0
 //
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
 #include "pch.h"
 
 #include "MaterialDefinition.h"
@@ -19,7 +23,7 @@ MaterialDefaultValues::MaterialDefaultValues(const UniformBufferDefinition& prop
 {
     AU_ASSERT(
         defaultProps.size() == propertyDefs.size(), "Default properties do not match definition");
-    for (int i = 0; i < defaultTxt.size(); i++)
+    for (size_t i = 0; i < defaultTxt.size(); i++)
     {
         textureNames.push_back(defaultTxt[i].name);
     }
